--- app/layout.module.scss
@@ -1,1 +1,0 @@
-@use 'utils' as *;
